+++ lib/has_crud/has_crud/action_controller/crud_admin_additions.rb
@@ -19,7 +19,7 @@ module HasCrud
                   :search_fields, :is_searchable?, :is_sortable?, :is_ajaxable?,
                   :is_exportable?, :title_for, :per_page, :settings_prefix
         base.send :respond_to, :html, :js, :csv
-        base.send :before_filter, :allow_all_parameters!
+        base.send :before_action, :allow_all_parameters!
       end
 
       module ClassMethods
@@ -67,7 +67,7 @@ module HasCrud
           end
 
           def update_resource(object, attributes)
-            result = object.update_attributes(*attributes)
+            result = object.update(*attributes)
             if result
               object.to_navigator! if object.respond_to? :to_navigator
             end